Performance fix.
Performace issue was caused by a lot of sphere colliders created by a CharacterController,
so I did a workaround (reduced CharacterController's collider size to minimal and added a
box collider to all characters on a screen). That reduced FPS drop
(till we have ~70 mobs on a screen or so, but it is much better now).